CommonFunctions: moved into separate library
