+++ saftcommon/CommonFunctions.cpp
