+++ src/saft-ctl.cpp
@@ -41,7 +41,7 @@
 #include "interfaces/SoftwareCondition.h"
 #include "interfaces/iDevice.h"
 #include "interfaces/iOwned.h"
-#include "src/CommonFunctions.h"
+#include "CommonFunctions.h"
 
 using namespace std;
