+++ src/saft-fg-ctl.cpp
@@ -33,7 +33,7 @@
 #include "interfaces/SCUbusCondition.h"
 #include "interfaces/FunctionGenerator.h" 
 #include "interfaces/FunctionGeneratorFirmware.h" 
-#include "src/CommonFunctions.h"
+#include "CommonFunctions.h"
 
 using namespace saftlib;
 using namespace std;
