+++ src/saft-gmt-check.cpp
@@ -42,7 +42,7 @@
 #include "interfaces/SoftwareCondition.h"
 #include "interfaces/iDevice.h"
 #include "interfaces/iOwned.h"
-#include "src/CommonFunctions.h"
+#include "CommonFunctions.h"
 
 using namespace std;
 using namespace saftlib;
